rag/requirements.txt
sentence-transformers==2.5.1
faiss-cpu==1.8.0
numpy==1.26.4
